requirements.txt
##############################
### Core Libraries & Frameworks
##############################
fastapi==0.115.12            # Async web framework (for building APIs)
uvicorn==0.34.2              # ASGI server (for running FastAPI apps)
Flask==3.1.1                 # Web framework (lightweight WSGI apps)
Werkzeug==3.1.3              # WSGI utility library (dependency for Flask)
Jinja2==3.1.6                # Template engine (dependency for Flask)
pymongo==4.13.0              # Official MongoDB driver
tortoise-orm==0.25.0         # Async ORM (supports PostgreSQL/MySQL, etc.)
pydantic==2.11.5             # Data validation library (core dependency of FastAPI)
markdown
bcrypt
psutil
# gunicorn                   # Only on Linux
waitress                     # WSGI service
apscheduler
tzlocal
openai
httpx
json_repair
backoff

playwright-stealth          # playwright more like brower
ultimate-sitemap-parser     # Sitemap

##############################
### Vector DB
##############################
hf_xet
chromadb==1.2.1
onnxruntime
# onnxruntime-gpu
sentence_transformers
langchain_text_splitters

##############################
### Network Requests & Parsing
##############################
requests==2.32.3             # HTTP requests library (includes SOCKS proxy support)
requests[socks]
urllib3==1.26.20             # HTTP client (Note: requires 1.26.9 ≤ ver < 1.27)
aiohttp==3.12.2              # Async HTTP client (dependency for FastAPI/Playwright)
beautifulsoup4==4.13.4       # HTML parsing library
bs4==0.0.2                   # Alias package for BeautifulSoup4
lxml==5.4.0                  # XML/HTML parser (high-performance XPath support)
feedparser==6.0.10           # RSS/Atom parser (version strictly pinned)
html2text==2025.4.15         # HTML to Markdown converter
markdownify==1.1.0           # HTML to Markdown converter
tldextract==5.3.0            # Domain suffix extraction tool
crawl4ai==0.6.3              # Web crawling SDK (specific version required)
playwright==1.52.0           # Install Browser automation:
                             #   set NODE_TLS_REJECT_UNAUTHORIZED=0
                             #   playwright install

##############################
### Data Processing & Storage
##############################
numpy==2.2.6                 # Numerical computation library (multidimensional array operations)
pandas==2.2.3                # Data analysis toolkit (tabular data processing)
faiss-cpu==1.11.0            # Vector similarity search (Windows compatible)
huggingface-hub==0.32.2      # Hugging Face model hub (includes `hf_xet` extension)

##############################
### Security & Encryption
##############################
cryptography==45.0.3         # Cryptography toolkit (TLS/SSL implementation)
pyOpenSSL==25.1.0            # OpenSSL wrapper (supports encrypted communication)
Brotli==1.1.0                # Compression algorithm (HTTP content encoding)
chardet==5.2.0               # Character encoding detection

##############################
### Async & Utilities
##############################
tenacity==9.1.2              # Retry mechanism library
watchdog==6.0.0              # Filesystem monitoring
PyQt5==5.15.11               # GUI development framework
typing_extensions==4.13.2    # Type hint extensions (Python compatibility)
pytest==8.3.5                # Testing framework (unit tests)
pytest-asyncio==1.0.0        # Async testing support
PyRSS2Gen==1.1               # RSS generator

##############################
### Commented Optional Dependencies
##############################
# faiss-gpu                   # GPU-accelerated version for Linux/Mac (requires NVIDIA env)
# faiss-cpu                   # Or use GPU version
# transformers==4.36.0        # Hugging Face NLP model library
# sentence-transformers       # Text embedding models (requires `transformers`)
# hnswlib                     # Approximate nearest neighbor search library
# PyQtWebEngine

streamlit
